wrapp agent with k8s node

+++ server/controller/src/main/java/ai/starwhale/mlops/domain/system/agent/bo/NodeInfo.java
@@ -29,6 +29,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class NodeInfo {
-    Integer memoryGB;
+    Float memoryGB;
     List<Device> devices;
 }
